# Significantly speed up parseString and other performance improvments