--- a/Sources/MaxMindDBSwift/GeoIP2.swift
+++ b/Sources/MaxMindDBSwift/GeoIP2.swift
@@ -186,35 +186,33 @@ public final class GeoIP2 {
     /// - Returns: Query result
     /// - Throws: GeoIP2Error if lookup fails
     public func lookup(ip: String) throws -> GeoIP2Result {
-        try queue.sync {
-            try ip.withCString { cString in
-                var gai_error: Int32 = 0
-                var mmdb_error: Int32 = 0
-                
-                // Execute query
-                let result = MMDB_lookup_string(mmdb, cString, &gai_error, &mmdb_error)
-                
-                // Check MMDB error
-                if mmdb_error != MMDB_SUCCESS {
-                    throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: nil)
-                }
-                
-                // Check network error
-                if gai_error != 0 {
-                    throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: gai_error)
-                }
-                
-                // If no entry is found, return empty result
-                guard result.found_entry else {
-                    return GeoIP2Result(data: [:])
-                }
-                
-                // Parse complete data
-                var entry = result.entry
-                let fullData = try parseFullData(entry: &entry)
-                
-                return GeoIP2Result(data: fullData)
+        try ip.withCString { cString in
+            var gai_error: Int32 = 0
+            var mmdb_error: Int32 = 0
+
+            // Execute query
+            let result = MMDB_lookup_string(mmdb, cString, &gai_error, &mmdb_error)
+
+            // Check MMDB error
+            if mmdb_error != MMDB_SUCCESS {
+                throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: nil)
             }
+
+            // Check network error
+            if gai_error != 0 {
+                throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: gai_error)
+            }
+
+            // If no entry is found, return empty result
+            guard result.found_entry else {
+                return GeoIP2Result(data: [:])
+            }
+
+            // Parse complete data
+            var entry = result.entry
+            let fullData = try parseFullData(entry: &entry)
+
+            return GeoIP2Result(data: fullData)
         }
     }
     
@@ -489,23 +487,21 @@ public final class GeoIP2 {
     /// - Returns: Database metadata dictionary
     /// - Throws: GeoIP2Error if retrieval fails
     public func metadata() throws -> [String: Any] {
-        return try queue.sync {
-            var entryList: UnsafeMutablePointer<MMDB_entry_data_list_s>?
-            let status = MMDB_get_metadata_as_entry_data_list(mmdb, &entryList)
-            
-            // Ensure resources are released
-            defer {
-                if let list = entryList {
-                    MMDB_free_entry_data_list(list)
-                }
+        var entryList: UnsafeMutablePointer<MMDB_entry_data_list_s>?
+        let status = MMDB_get_metadata_as_entry_data_list(mmdb, &entryList)
+
+        // Ensure resources are released
+        defer {
+            if let list = entryList {
+                MMDB_free_entry_data_list(list)
             }
-            
-            guard status == MMDB_SUCCESS, let list = entryList else {
-                throw GeoIP2Error.dataParsingFailed(reason: "Failed to get metadata")
-            }
-            
-            return try parseEntryDataList(entryList: list)
         }
+
+        guard status == MMDB_SUCCESS, let list = entryList else {
+            throw GeoIP2Error.dataParsingFailed(reason: "Failed to get metadata")
+        }
+
+        return try parseEntryDataList(entryList: list)
     }
     
     /// Return data directly as a JSON string
@@ -525,57 +521,55 @@ public final class GeoIP2 {
     /// - Returns: Raw data JSON string
     /// - Throws: Error if lookup fails
     public func getRawDataJSON(ip: String) throws -> String {
-        return try queue.sync {
-            try ip.withCString { cString in
-                var gai_error: Int32 = 0
-                var mmdb_error: Int32 = 0
-                
-                // Execute query
-                let result = MMDB_lookup_string(mmdb, cString, &gai_error, &mmdb_error)
-                
-                // Check errors
-                if mmdb_error != MMDB_SUCCESS {
-                    throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: nil)
-                }
-                if gai_error != 0 {
-                    throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: gai_error)
-                }
-                
-                // If no entry is found, return empty result
-                guard result.found_entry else {
-                    return "{}"
-                }
-                
-                var entry = result.entry
-                var entryList: UnsafeMutablePointer<MMDB_entry_data_list_s>?
-                let status = MMDB_get_entry_data_list(&entry, &entryList)
-                
-                // Ensure resources are released
-                defer {
-                    if let list = entryList {
-                        MMDB_free_entry_data_list(list)
-                    }
-                }
-                
-                guard status == MMDB_SUCCESS, let list = entryList else {
-                    throw GeoIP2Error.dataParsingFailed(reason: "Failed to get entry data list")
-                }
-                
-                // Parse data
-                let data = try parseEntryDataList(entryList: list)
-                
-                // Convert to JSON
-                let options: JSONSerialization.WritingOptions = [.prettyPrinted, .sortedKeys]
-                guard let jsonData = try? JSONSerialization.data(withJSONObject: data, options: options),
-                      let jsonString = String(data: jsonData, encoding: .utf8) else {
-                    throw GeoIP2Error.dataParsingFailed(reason: "Failed to convert data to JSON")
-                }
-                
-                return jsonString
+        try ip.withCString { cString in
+            var gai_error: Int32 = 0
+            var mmdb_error: Int32 = 0
+
+            // Execute query
+            let result = MMDB_lookup_string(mmdb, cString, &gai_error, &mmdb_error)
+
+            // Check errors
+            if mmdb_error != MMDB_SUCCESS {
+                throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: nil)
             }
+            if gai_error != 0 {
+                throw GeoIP2Error.lookupFailed(code: mmdb_error, gaiError: gai_error)
+            }
+
+            // If no entry is found, return empty result
+            guard result.found_entry else {
+                return "{}"
+            }
+
+            var entry = result.entry
+            var entryList: UnsafeMutablePointer<MMDB_entry_data_list_s>?
+            let status = MMDB_get_entry_data_list(&entry, &entryList)
+
+            // Ensure resources are released
+            defer {
+                if let list = entryList {
+                    MMDB_free_entry_data_list(list)
+                }
+            }
+
+            guard status == MMDB_SUCCESS, let list = entryList else {
+                throw GeoIP2Error.dataParsingFailed(reason: "Failed to get entry data list")
+            }
+
+            // Parse data
+            let data = try parseEntryDataList(entryList: list)
+
+            // Convert to JSON
+            let options: JSONSerialization.WritingOptions = [.prettyPrinted, .sortedKeys]
+            guard let jsonData = try? JSONSerialization.data(withJSONObject: data, options: options),
+                  let jsonString = String(data: jsonData, encoding: .utf8) else {
+                throw GeoIP2Error.dataParsingFailed(reason: "Failed to convert data to JSON")
+            }
+
+            return jsonString
         }
     }
-    
+
     /// Parse MAP type data
     /// - Parameter entryList: Data list pointer
     /// - Returns: Parsed dictionary