I profiled this code and found most of the time was spent in
`parseString`. This PR improves its performance significantly.
The results of my benchmarking are below.

## Before

```
name         time          std        iterations
------------------------------------------------
Lookup IP     67871.000 ns ±   1.58 %      20509
Lookup JSON  154864.000 ns ±   1.24 %       9023
Get Raw JSON 214237.000 ns ±   0.96 %       6488
```

![image](https://github.com/user-attachments/assets/aee6b730-7907-47b2-8559-b80f54738517)


## After refactoring `parseString` but before removing caching

```
name         time          std        iterations
------------------------------------------------
Lookup IP     33787.000 ns ±   2.02 %      41122
Lookup JSON  123664.000 ns ±   1.45 %      11291
Get Raw JSON 170356.000 ns ±   1.23 %       7835
```

## After removing caching

```
name         time          std        iterations
------------------------------------------------
Lookup IP      3886.000 ns ±   4.78 %     357934
Lookup JSON   91716.000 ns ±   2.20 %      15520
Get Raw JSON 138819.000 ns ±   4.73 %      10072
```

![image](https://github.com/user-attachments/assets/f9e108ad-c825-4a53-b9f0-22f917546c8f)