@oschwald 2025-05-01
Actually, I am finding the caching is actually slowing down the lookup, after the improvements. I am going to drop that and fix some related issues.