@oschwald 2025-05-01
Here is the performance after dropping the caching:

name         time          std        iterations
------------------------------------------------
Lookup IP      3886.000 ns ±   4.78 %     357934
Lookup JSON   91716.000 ns ±   2.20 %      15520
Get Raw JSON 138819.000 ns ±   4.73 %      10072

If we do want to retain the caching for memory reasons, perhaps we can find a way to better optimize it.